Added poses of running/staying with/without weapon.
Added function when player is holding weapon default pose changes on holding gun player pose, otherwise it changes on nonholding gun player pose.

UPD: It causes a bug with running animation. When the poses are changing the object that was turned off are turning off the animations too so animation remembers last state and after that it starts from this state all the time.