Add punch system.
Added punch functionality if player without gun. The thing is the way i solve a problem is not too good and i think it may cause problems with optimization in future. Seems it can be a problem in "Punching" script. However it works.

Also i fixed a bud with picking up one gun is lying near the other.