Finally fixed a problem with shooting through a wall.
Added players collider when i checking up an object in front of player if the weapon on. Maybe it need to be solved something else, like adding a new child object to weapon with tag "Player" and box collider to check collisions. BUT........did what i did. Making a commit inside train. :D